more bug fixes. needs testing and then upload update.

+++ ProgrammingIdeas/Helpers/CloudDB.cs
@@ -13,10 +13,11 @@ using System.Linq;
 using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
+using Android.Support.V7.Widget;
 
 namespace ProgrammingIdeas.Helpers
 {
-    internal class CloudDB
+    internal static class CloudDB
     {
         private static string oldDBPath = Path.Combine(Global.APP_PATH, "ideasdb");
         private static string newideastxtPath = Path.Combine(Global.APP_PATH, "newideastxt");
@@ -26,19 +27,15 @@ namespace ProgrammingIdeas.Helpers
         private static Activity activity;
         private const string TAG = "ALANSADEBUG";
 
-        public static void Init(Activity activity)
-        {
-            CloudDB.activity = activity;
-        }
-
         public static async Task Startup(Action retryMethod, Snackbar snack)
         {
             try
             {
-                var source = new CancellationTokenSource();
-                source.CancelAfter(12000);
+				activity = App.CurrentActivity;
                 if (!File.Exists(oldDBPath)) // First launch, as in just installed app and launched it
                 {
+					var source = new CancellationTokenSource();
+					source.CancelAfter(12000);
                     var response = await client.GetAsync(AppResources.DbLink, source.Token);
                     if (response.IsSuccessStatusCode)
                     {
@@ -76,22 +73,22 @@ namespace ProgrammingIdeas.Helpers
 
         private static void Notify(int newIdeasCount)
         {
-            activity.RunOnUiThread(() =>
-            {
-                string notifContent = newIdeasCount == 1 ? $"1 new idea is available." : $"{newIdeasCount.ToString()} new ideas are available.";
-                var intent = new Intent(activity, typeof(CategoryActivity));
-                intent.PutExtra("NewIdeasNotif", true);
-                var pendingIntent = PendingIntent.GetActivity(activity, 1960, intent, PendingIntentFlags.UpdateCurrent);
+			activity.RunOnUiThread(() =>
+			{
+				string notifContent = newIdeasCount == 1 ? $"1 new idea is available." : $"{newIdeasCount.ToString()} new ideas are available.";
+				var intent = new Intent(activity, typeof(CategoryActivity));
+				intent.PutExtra("NewIdeasNotif", true);
+				var pendingIntent = PendingIntent.GetActivity(activity, 1960, intent, PendingIntentFlags.UpdateCurrent);
 
-                var notif = new NotificationCompat.Builder(activity)
-                                                  .SetContentTitle("New ideas available.")
-                                                  .SetContentText(notifContent)
-                                                  .SetContentIntent(pendingIntent)
-                                                  .SetSmallIcon(Resource.Mipmap.notif_icon)
-                                                  .Build();
-                var notifManager = (NotificationManager)activity.GetSystemService(Context.NotificationService);
-                notifManager.Notify(1957, notif);
-            })
+				var notif = new NotificationCompat.Builder(activity)
+												  .SetContentTitle("New ideas available.")
+												  .SetContentText(notifContent)
+												  .SetContentIntent(pendingIntent)
+												  .SetSmallIcon(Resource.Mipmap.notif_icon)
+												  .Build();
+				var notifManager = (NotificationManager)activity.GetSystemService(Context.NotificationService);
+				notifManager.Notify(1957, notif);
+			});
         }
 
         private static bool NewIdeasAvailable(List<Category> oldIdeas, List<Category> newIdeas)
@@ -132,20 +129,25 @@ namespace ProgrammingIdeas.Helpers
                             Log.Debug(TAG, "Starting full invalidation");
 
                             var notesPath = Path.Combine(Global.APP_PATH, "notesdb");
-                            File.Create(notesPath);
+							if(!File.Exists(notesPath))
+								File.Create(notesPath);
                             var notes = JsonConvert.DeserializeObject<List<Note>>(DBAssist.DeserializeDB(notesPath));
                             notes = notes ?? new List<Note>();
-                            for (int i = 0; i < Global.Categories.Count; i++)
+							for (int i = 0; i < newDB.Count; i++)
                             {
-                                for (int j = 0; j < Global.Categories[i].Items.Count; j++)
+                                for (int j = 0; j < newDB[i].Items.Count; j++)
                                 {
                                     var newItem = newDB[i].Items[j];
                                     var oldItem = Global.Categories[i].Items.FirstOrDefault(x => x.Id == newItem.Id);
                                     Note note = null;
-                                    if (oldItem != null)
-                                        note = notes.FirstOrDefault(x => x.Title == oldItem.Title);
+									if (oldItem != null)
+									{
+										note = notes.FirstOrDefault(x => x.Title == oldItem.Title);
+										if(note != null)
+											Log.Debug(TAG, $"Note *{note.Title}*, found for old idea *{oldItem.Title}* placed at new idea *{newItem.Title}*");
+									}
 
-                                    newItem.Note = note ?? note;
+                                    newItem.Note = note;
                                     newItem.State = oldItem?.State;
                                 }
                             }
